update local experiments
